fix: readonly/writeonly protocol resource handling
