+++ bevy_rtc/src/client/mod.rs
@@ -1,12 +1,13 @@
+#[allow(clippy::module_inception)]
+mod client;
 mod events;
 mod plugin;
 mod router;
 mod state;
-mod system_params;
 mod systems;
 
+pub use client::RtcClient;
 pub use events::{RtcClientEvent, RtcClientRequestEvent};
 pub use plugin::RtcClientPlugin;
 pub use router::AddClientProtocolExt;
 pub use state::{RtcClientState, RtcClientStatus};
-pub use system_params::RtcClient;
