+++ bevy_rtc/src/server/mod.rs
@@ -1,12 +1,12 @@
+mod client;
 mod events;
 mod plugin;
 mod router;
 mod state;
-mod system_params;
 mod systems;
 
+pub use client::RtcServer;
 pub use events::RtcServerEvent;
 pub use plugin::RtcServerPlugin;
 pub use router::AddServerProtocolExt;
 pub use state::{RtcServerState, RtcServerStatus};
-pub use system_params::RtcServer;
